Get rid of singletons in Core modules
Core "Managers" are now called "Systems", to follow more in line with
what part of LukeBot they are.

Systems are no longer done with Singleton design and are initialized as
static fields of Systems class. This is mostly to have more control over
their initialization.

Other parts were adjusted to work with these changes.