Core: Introduce EventSystem
Current "event" system (or lack thereof) was mostly giving event
listeners direct access to event publishers and use simple and "classic"
C# event mechanisms. This, however, quickly fell apart with introduction
of PubSub - if there is a listener which wants to listen to two
publishers, we have to provide them both, which unnecessarily
complicates and spaghetti-ies the code.

EventSystem is meant to be a middle-man between publishers and listeners.
Its main goal is to create a sort of "abstraction layer" between these
two elements and organize the code, providing freedom to subscribe to
events and register as publishers.

There is no support for custom events yet but this system will make it
way easier to add such support if need be.