Core: Implement Property Store
Property Store is meant to serve as a central point for various
properties connected to LukeBot's settings. Should be accessible
from Core to other parts of LukeBot to pull various settings.

Property Store was added to tests for now, will be properly used later
together with a tool to modify properties on the fly.