Widget: Fix race condition on AcquireWS
When attempting to provide a new WebSocket connection to already
connected Widget WebSocket (ex. when pressing "Refresh" button on OBS
Browser Widget) there was a race condition where existing WebSocket
would be overwritten by a new one. Additional bonus was that Event
marking the end of WebSocket lifetime was not reset, which caused new
connection to end immediately and new WS to be disposed immediately as
well.

LifetimeEndEvent is now set in CloseWS() to prevent duplicating code
unnecessarily. AcquireWS() will also ensure that the LifetimeEndEvent
is Reset before forming a LifetimeTask.

Additionally "ref" keywords were removed when passing WS from Endpoint
to Widget - figured it's not really necessary, as objects are references
anyway...