Property Store: Add "hidden" property attribute
This is meant mostly to hide Properties from displaying them on screen
in some insecure situations (ex. during stream). This parameter will
skip displaying the Properties when using propmgr.

Note that this does not hide the Properties from code perspective. It
only is meant as a hiding point like files starting with . on Unix
systems.

To always show hidden properties use "propmgr list -a".