CLI: Change bools to unified State variable
This changes the CLI to actually manage its own state via a State
member. Thanks to that we know exactly in which state we were put by
the user or log events and can react accordingly.

This change also fixes a deadlock which could happen when another thread
tries to log a message while we are in the middle of processing a CLI command.