API: Ensure Token is not expired
This adds bits of code to make sure Token has not expired yet. If it
did, API module will automatically refresh it to prevent a potential
Unauthorized message.

API still will attempt another refresh if we receive an Unauthorized
response as it used to, but now that code path should be way less
frequently taken.

AuthorizationCodeFlow also now uses a randomly generated state variable
instead of the trusty ol' CHANGEMETOSOMETHINGRANDOM