EventSubClient: Allow for more graceful error handling
If ReceiveAsync() ever throws an exception, now it will be captured and
a reconnect attempt will be made. If reconnection attempt throws, that
will also be captured and the thread will exit gracefully without
breaking the rest of the bot.

This still needs some extra code to ex. manually restart specific
services (like EventSubClient's thread) but should at least make it more
error-proof.