EventSystem: Fix returned event callbacks
Somehow, amidst all the work and code to rework the Event System, I
forgot to change this bit...

Event publishers would receive their own callback to publish an event.
However, as it turned out when a subscription crashed my bot during my
last stream, the event still executed on publisher's thread. This change
fixes that by actually submitting an event to the Dispatcher instead of
raising it directly.

I should write more tests for this code...