Twitch: Add Badge support to chat
Available Badges are fetched from Twitch's API and then forwarded when
needed to a Chat Widget when chat message occurs.

Twitch's CDN is a bit finnicky sometimes and it failed to load the
images at first attempt. To help that some code was added to retry it up
to 5 times. Once the badge images are properly loaded OBS browser plugin
should cache the results and no more problems like that should occur.