Twitch: Add external emote refresh
This allows to trigger an update of external emote collection (coming
from FFZ, BTTV and/or 7TV), which reflects changes to emote sets done on
these services.

It is difficult to make this process automatic. So far the only service
which somewhat allows to automate this process is 7TV, and it is not
documented whatsoever. Performing a manual trigger like that would still
be necessary for other services, so manually triggering a reload of all
emote sets seems the most reasonable idea.

Trigger the reload using "twitch emote-refresh" CLI command.