Refactor Module/Service library
This is a groundwork for having this project be more organized.

In short - LukeBot.Module is now named LukeBot.Services. It also adds
IService, which is an interface that other LukeBot services have to
implement.

To hide each service's implementation, each creates its own I*Service
interface which implements/extends IService. Various part of LukeBot
should use I*Service interfaces to communicate with Bot's services.

LukeBot.Services.Service now forgoes "static" services. Instead, each
Service instance is created and added on runtime when starting the Bot.
To access each Service there is a Service.Get(string) call.

There probably is more, but I worked on it on and off for a long while
so it's time to make a checkpoint here.

At this state Services load, but no User Modules are loaded yet. That is
to be solved soon.