WidgetConfiguration: Support restricted fields
Adds WidgetConfigurationRestrictedField generic attribute which can be
used to restrict field's possible values.

Attribute accepts a parameter - type of validator class that is going to
be used as validation code when calling the Field's Setter. Naturally,
this can only work if the Field is set through the Field class and not
directly.

Tests were added to check this behavior.