Client: Delegate Query processing to separate thread
Whenever Client receives a Query message its response will now be
processed on a separate CommandExecutor. This is to free the Receive
thread from waiting for user input and to allow processing other
commands.

This fixes server-client Ping functionality when being in the middle of
processing a long Query-filled command (ex. with Widget configuration
editor).